Fixes case in which config.when() and transform work together
There's some subtlety with the name. Previously,
config.when would change the function name accordingly. This
was so that later on decorators could refer to the function by
its true name. However, this broke depending on decorator ordering.
This change adds a sanitize_name functionality to unblock the transform decorator,
but does not yet take away the name changing functionality. We'll need to think through
that, but this is a happy medium for now I think.