Adds ability to change the return type of Hamilton execution
Hamilton was hardcoded to return a pandas dataframe. This change
introduces a module `base` that houses:

1. an interface object `ResultMixin` with a static function called "build_result".
2. an interface object `HamiltonExecutor` that houses the required interface to implement execution.

`ResultMixin`: The idea is that we'll use this class as the interface to describe the shape required to return different results.
I then went ahead and implemented DictResult, PandasDataFrameResult, and NumpyMatrixResult.
I still have to add some unit tests for them.

All executors will inherit and override the functions in `HamiltonExecutor`. I decided to have it extend ResultMixin to simplify
things. Otherwise I include the `SimplePythonDataFrameExecutor`, which mirrors the current Hamilton
execution behavior.