Refactors Hamilton DaskExecutor and unit test
1. We can never have an internal package that matches the name
of a top level package we want to use -- it just gets confusing. So renamed
dask module in hamilton to h_dask to differentiate.
2. Fixes up code to run -- the unit test passes when I run it locally.

The changes to the DaskExecutor involve adding a constructor
and the ability to change the result type returned. Dask is general
purpose so we should expose this as such.

Dask also has a visualization component, so wire through the ability
to see the graph as well. Note, we should probably divorce execution
and visualization --- but right now when iterating this is the simplest
amount of code to maintain.