Adds example way to run Dask
This adds a hello world dask example.

One key idea is that data loading should be abstracted
into hamilton functions. That way, if you want to change
how things are loaded, you just change the module imported.

I add the output visualizations because that will
be useful to assess how things change if this example is
updated/changed.